nat: extract DNAT protocol/proto-resolver tests into tests_dnat_proto.rs (#4409)
Move the 9 DNAT protocol-match tests -- dnat_protocol_{gre,icmp6}_
translates, dnat_ip_only_covers_all_protocols_incl_icmp,
dnat_concrete_rule_wins_over_ip_only_wildcard,
dnat_invalid_destination_address_yields_no_entry,
dnat_protocol_zero_hopopt_is_distinct_from_wildcard and the three
proto_number_* resolver tests -- verbatim into tests_dnat_proto.rs. No
test-local helpers.

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 32 tests.
