+++ userspace-dp/src/nat/mod.rs
@@ -47,6 +47,9 @@ mod tests_pool;
 #[cfg(test)]
 #[path = "tests_counter.rs"]
 mod tests_counter;
+#[cfg(test)]
+#[path = "tests_dnat_proto.rs"]
+mod tests_dnat_proto;
 
 /// #3096: per-flow interface / routing-instance identity passed into the NAT
 /// match path so an interface- or routing-instance-scoped rule-set matches
